settings container + layout
