+++ deploy-web/src/app/settings/SettingsForm.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { useState, useRef } from "react";
 import {
   Box,
