+++ deploy-web/src/app/settings/page.tsx
@@ -1,0 +1,10 @@
+import { Metadata } from "next";
+import { SettingsContainer } from "./SettingsContainer";
+
+export const metadata: Metadata = {
+  title: "Settings"
+};
+
+export default function Settings() {
+  return <SettingsContainer />;
+}
